Refactor:예외 처리 수정
- 게시글,댓글에 대한 예외처리 수정
